% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allom_pop.R
\name{allom_pop}
\alias{allom_pop}
\title{Allometric adjustment of morphometric characters to correct for
ontogenetic variation in a multipopulation (single species) dataset}
\usage{
allom_pop(data)
}
\arguments{
\item{data}{csv file with population identifiers in the 1st column, followed by body length measurement (e.g. snout-vent-length) in the 2nd column. Other morphometric measurements are contained in the 3rd column onwards. Singletons (only one sample per species) and missing data are not allowed}
}
\value{
Returns log-transformed and body-size adjusted data in a table called outfile.csv
}
\description{
This function uses the following allometric growth equation to adjust for ontogenetic variation:
Xadj = log10(X)-B[log10(SVL)-log10(SVLmean)], where Xadj = Adjusted value for character X;
X = raw/unadjusted value for character X; B = pooled regression coefficient (slope) of log10(X) against log10(SVL);
SVL = measured SVL; SVLmean = grand mean SVL for a particular species
}
\details{
Following Thorpe(1975), this function uses common within-group pooling to obtain a separate slope for each population.
SVLmean is the grand mean calculated by averaging over all populations. 
Each species should be represented by more than two individuals for the adjustment to work and missing data must not be included. 
This adjustment should also be performed separately on different sexes to account for possible sexual dimorphism.
}
\examples{
m <- read.csv("foo.csv") ## foo.csv = raw measurements
allom_pop(m)
}
\references{
Thorpe, R. S. (1975) Quantitative handling of characters useful in snake systematics with particular reference to
intraspecific variation in the Ringed Snake Natrix natrix (L.). Biological Journal of the Linnaean Society, 7: 27-43
}
