% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_adjust.R
\name{body_adjust}
\alias{body_adjust}
\title{Allometric adjustment of morphometric characters to correct for intraspecific
ontogenetic variation.}
\usage{
body_adjust(data)
}
\arguments{
\item{data}{Path to the input file}
}
\value{
Returns log-transformed and adjusted data
}
\description{
This function assumes that each Operational Taxonomic
Unit represents a distinct species. The input file is data frame in csv format. The firt column contains OTU identifiers, followed by
morphological characters. No other data should be included in the data frame.
Singleton OTU's and juvenile measurements should be exlucded.
}
