% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{ez_pca}
\alias{ez_pca}
\title{Performs PCA}
\usage{
ez_pca(data)
}
\arguments{
\item{data}{csv file with OTU identifiers in the first column, followed by morphological characters, each in in a separate column.}
}
\value{
Returns a summary table called PCA_summary.csv that includes eigenvalues and PCA loadings in addition
 to PCA plots of the first three PC's
}
\description{
Performs PCA using the base function prcomp with scaling. Outputs ggplot graphs
and a summary table
}
