% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_struct.R
\name{morpho_struct}
\alias{morpho_struct}
\title{Perform allometric size correction, PCA, and ANOVA/TUKEY analyses on morphological data}
\usage{
morpho_struct(
  input_file,
  type = "species",
  colors = NULL,
  draw_ellipses = TRUE,
  confidence_level = 0.95
)
}
\arguments{
\item{input_file}{Raw morphological data in CSV format. See `allom()` documention on how to format input data}

\item{type}{Type of allometric correction to use. Choose from "species", "population1" or "population2". See `allom()` documention}

\item{colors}{Optional vector of colors for the PCA plot. Number of colors need to equal number of groups/species}

\item{draw_ellipses}{Boolean. Whether to draw confidence ellipse on PCA plot}

\item{confidence_level}{Level of confidence interval to draw ellipse}
}
\value{
PCA plot of PC1 and PC2 scores. Also writes two tables. One containing a summary of the PCA analysis and another containing the results of the ANOVA/Tukey analysis
}
\description{
Wrapper function that takes raw morphological data, performs allometric size correction using `allom()`, followed by PCA and ANOVA-Tukey test on size-corrected data.
}
\examples{
Path to raw morphological data
mydata <- "path/to/morpho_data.csv"

morpho_struct("mydata", type = "species")
}
