% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdist.R
\name{plot_pdist}
\alias{plot_pdist}
\title{Create boxplots comparing p-distances of a focal species vs. a list of other species}
\usage{
plot_pdist(raw_distance_df, focal_species, comparison_species)
}
\arguments{
\item{focal_species}{Focal species that you want to make comparisons with}

\item{comparison_species}{A list of other species to be compared with the focal species}

\item{x}{Output from the pdist_cutoff function}
}
\value{
Writes two CSV-formatted files. One contains raw p-distances for every individual pairwise comparison. A second table aggregates p-distances by species showing min, max, and mean. Species that meet the cut-off will be printed
}
\description{
Uses output from the pdist_cutoff function
}
\examples{
Set path to fasta alignment
myfasta <- "path/to/fasta/fasta_alignment.fas"

Set cutoff value
mycutoff <- 0.05

Set your output file name
raw_output <- "raw_output.csv"
aggregated_output <- "aggregated_output.csv"

Get p-distances
results <- pdist_cutoff(myfasta, mycutoff, raw_output, aggregated_output)

Specify the focal species  and the list of comparison species
focal_species <- "Micryletta_subaraji"
comparison_species <- c("Micryletta_sumatrana", "Micryletta_dissimulans", "Micryletta_inornata", "Micryletta_nigromaculata")

Boxplots for the specified species combinations
plot_pdist(results, focal_species, comparison_species)
}
