% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid}
\alias{resid}
\title{Calculate residuals of a linear regression of each character against body size}
\usage{
resid(data)
}
\arguments{
\item{data}{csv file with species identifiers in the 1st column, followed by body length measurement (e.g. snout-vent-length) in the 2nd column. Other morphometric measurements are contained in the 3rd column onwards. Singletons (only one sample per species) and missing data are not allowed}
}
\value{
Returns residuals (slope) for each character
}
\description{
Calculate residuals of a linear regression of each character against body size
}
\examples{
m <- read.csv("foo.csv") ## foo.csv = raw measurements
resid(m)
}
