% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist_cutoff.R
\name{pdist_cutoff}
\alias{pdist_cutoff}
\title{Calculate p-distances and identify species pairs that are above a user-specified threshold}
\usage{
pdist_cutoff(fasta_file, cutoff, output_file)
}
\arguments{
\item{fasta_file}{Alignned FASTA sequences. Make sure sequences are labeled as genus_species. You can add other unique identifiers after the species name using a second underscore. E.g. Bufo_bufo_MN09667}

\item{cutoff}{Set your desired cut-off. E.g. 0.03 for a 3 percent cut-off}

\item{output_file}{Name of output file}
}
\value{
Returns a CSV-formatted file with species pairs in separate columns followed by min, max, and mean
}
\description{
Uses the dist_dna function in ape to calculate p-distances
}
\examples{
Set path to fasta alignment
myfasta <- "path/to/fasta/fasta_alignment.fas"

Set cutoff value
mycutoff <- 0.05

Set your output file name
myoutput <- "pdist_output.csv"

Run example
pdist_cutoff(myfasta, mycutoff, myoutput)
}
